Pull contracts into own file
