+++ packages/multi-test/src/lib.rs
@@ -1,9 +1,11 @@
 mod app;
 mod bank;
+mod contracts;
 mod test_helpers;
 mod transactions;
 mod wasm;
 
 pub use crate::app::{parse_contract_addr, App, AppCache, AppOps};
 pub use crate::bank::{Bank, BankCache, BankOps, SimpleBank};
-pub use crate::wasm::{next_block, Contract, ContractWrapper, WasmCache, WasmOps, WasmRouter};
+pub use crate::contracts::{Contract, ContractWrapper};
+pub use crate::wasm::{next_block, WasmCache, WasmOps, WasmRouter};
