+++ packages/multi-test/src/wasm.rs
@@ -1,283 +1,15 @@
 use schemars::JsonSchema;
-use serde::de::DeserializeOwned;
 use std::collections::HashMap;
 use std::fmt;
 use std::ops::Deref;
 
 use cosmwasm_std::{
-    from_slice, Addr, Api, Binary, BlockInfo, ContractInfo, CosmosMsg, Deps, DepsMut, Empty, Env,
-    MessageInfo, Querier, QuerierWrapper, Reply, Response, Storage, SubMsg, WasmQuery,
+    Addr, Api, Binary, BlockInfo, ContractInfo, Deps, DepsMut, Env, MessageInfo, Querier,
+    QuerierWrapper, Reply, Response, Storage, WasmQuery,
 };
 
+use crate::contracts::Contract;
 use crate::transactions::{RepLog, StorageTransaction};
-
-/// Interface to call into a Contract
-pub trait Contract<T>
-where
-    T: Clone + fmt::Debug + PartialEq + JsonSchema,
-{
-    fn execute(
-        &self,
-        deps: DepsMut,
-        env: Env,
-        info: MessageInfo,
-        msg: Vec<u8>,
-    ) -> Result<Response<T>, String>;
-
-    fn instantiate(
-        &self,
-        deps: DepsMut,
-        env: Env,
-        info: MessageInfo,
-        msg: Vec<u8>,
-    ) -> Result<Response<T>, String>;
-
-    fn sudo(&self, deps: DepsMut, env: Env, msg: Vec<u8>) -> Result<Response<T>, String>;
-
-    fn reply(&self, deps: DepsMut, env: Env, msg: Reply) -> Result<Response<T>, String>;
-
-    fn query(&self, deps: Deps, env: Env, msg: Vec<u8>) -> Result<Binary, String>;
-}
-
-type ContractFn<T, C, E> =
-    fn(deps: DepsMut, env: Env, info: MessageInfo, msg: T) -> Result<Response<C>, E>;
-type SudoFn<T, C, E> = fn(deps: DepsMut, env: Env, msg: T) -> Result<Response<C>, E>;
-type ReplyFn<C, E> = fn(deps: DepsMut, env: Env, msg: Reply) -> Result<Response<C>, E>;
-type QueryFn<T, E> = fn(deps: Deps, env: Env, msg: T) -> Result<Binary, E>;
-
-type ContractClosure<T, C, E> = Box<dyn Fn(DepsMut, Env, MessageInfo, T) -> Result<Response<C>, E>>;
-type SudoClosure<T, C, E> = Box<dyn Fn(DepsMut, Env, T) -> Result<Response<C>, E>>;
-type ReplyClosure<C, E> = Box<dyn Fn(DepsMut, Env, Reply) -> Result<Response<C>, E>>;
-type QueryClosure<T, E> = Box<dyn Fn(Deps, Env, T) -> Result<Binary, E>>;
-
-/// Wraps the exported functions from a contract and provides the normalized format
-/// Place T4 and E4 at the end, as we just want default placeholders for most contracts that don't have sudo
-pub struct ContractWrapper<T1, T2, T3, E1, E2, E3, C = Empty, T4 = Empty, E4 = String, E5 = String>
-where
-    T1: DeserializeOwned,
-    T2: DeserializeOwned,
-    T3: DeserializeOwned,
-    T4: DeserializeOwned,
-    E1: ToString,
-    E2: ToString,
-    E3: ToString,
-    E4: ToString,
-    E5: ToString,
-    C: Clone + fmt::Debug + PartialEq + JsonSchema,
-{
-    execute_fn: ContractClosure<T1, C, E1>,
-    instantiate_fn: ContractClosure<T2, C, E2>,
-    query_fn: QueryClosure<T3, E3>,
-    sudo_fn: Option<SudoClosure<T4, C, E4>>,
-    reply_fn: Option<ReplyClosure<C, E5>>,
-}
-
-impl<T1, T2, T3, E1, E2, E3, C> ContractWrapper<T1, T2, T3, E1, E2, E3, C>
-where
-    T1: DeserializeOwned + 'static,
-    T2: DeserializeOwned + 'static,
-    T3: DeserializeOwned + 'static,
-    E1: ToString + 'static,
-    E2: ToString + 'static,
-    E3: ToString + 'static,
-    C: Clone + fmt::Debug + PartialEq + JsonSchema + 'static,
-{
-    pub fn new(
-        execute_fn: ContractFn<T1, C, E1>,
-        instantiate_fn: ContractFn<T2, C, E2>,
-        query_fn: QueryFn<T3, E3>,
-    ) -> Self {
-        ContractWrapper {
-            execute_fn: Box::new(execute_fn),
-            instantiate_fn: Box::new(instantiate_fn),
-            query_fn: Box::new(query_fn),
-            sudo_fn: None,
-            reply_fn: None,
-        }
-    }
-
-    /// this will take a contract that returns Response<Empty> and will "upgrade" it
-    /// to Response<C> if needed to be compatible with a chain-specific extension
-    pub fn new_with_empty(
-        execute_fn: ContractFn<T1, Empty, E1>,
-        instantiate_fn: ContractFn<T2, Empty, E2>,
-        query_fn: QueryFn<T3, E3>,
-    ) -> Self {
-        ContractWrapper {
-            execute_fn: customize_fn(execute_fn),
-            instantiate_fn: customize_fn(instantiate_fn),
-            query_fn: Box::new(query_fn),
-            sudo_fn: None,
-            reply_fn: None,
-        }
-    }
-}
-
-impl<T1, T2, T3, E1, E2, E3, C, T4, E4> ContractWrapper<T1, T2, T3, E1, E2, E3, C, T4, E4>
-where
-    T1: DeserializeOwned + 'static,
-    T2: DeserializeOwned + 'static,
-    T3: DeserializeOwned + 'static,
-    T4: DeserializeOwned + 'static,
-    E1: ToString + 'static,
-    E2: ToString + 'static,
-    E3: ToString + 'static,
-    E4: ToString + 'static,
-    C: Clone + fmt::Debug + PartialEq + JsonSchema + 'static,
-{
-    pub fn new_with_sudo(
-        execute_fn: ContractFn<T1, C, E1>,
-        instantiate_fn: ContractFn<T2, C, E2>,
-        query_fn: QueryFn<T3, E3>,
-        sudo_fn: SudoFn<T4, C, E4>,
-    ) -> Self {
-        ContractWrapper {
-            execute_fn: Box::new(execute_fn),
-            instantiate_fn: Box::new(instantiate_fn),
-            query_fn: Box::new(query_fn),
-            sudo_fn: Some(Box::new(sudo_fn)),
-            reply_fn: None,
-        }
-    }
-}
-
-impl<T1, T2, T3, E1, E2, E3, C, E5> ContractWrapper<T1, T2, T3, E1, E2, E3, C, Empty, String, E5>
-where
-    T1: DeserializeOwned + 'static,
-    T2: DeserializeOwned + 'static,
-    T3: DeserializeOwned + 'static,
-    E1: ToString + 'static,
-    E2: ToString + 'static,
-    E3: ToString + 'static,
-    E5: ToString + 'static,
-    C: Clone + fmt::Debug + PartialEq + JsonSchema + 'static,
-{
-    pub fn new_with_reply(
-        execute_fn: ContractFn<T1, C, E1>,
-        instantiate_fn: ContractFn<T2, C, E2>,
-        query_fn: QueryFn<T3, E3>,
-        reply_fn: ReplyFn<C, E5>,
-    ) -> Self {
-        ContractWrapper {
-            execute_fn: Box::new(execute_fn),
-            instantiate_fn: Box::new(instantiate_fn),
-            query_fn: Box::new(query_fn),
-            sudo_fn: None,
-            reply_fn: Some(Box::new(reply_fn)),
-        }
-    }
-}
-
-fn customize_fn<T, C, E>(raw_fn: ContractFn<T, Empty, E>) -> ContractClosure<T, C, E>
-where
-    T: DeserializeOwned + 'static,
-    E: ToString + 'static,
-    C: Clone + fmt::Debug + PartialEq + JsonSchema + 'static,
-{
-    let customized =
-        move |deps: DepsMut, env: Env, info: MessageInfo, msg: T| -> Result<Response<C>, E> {
-            raw_fn(deps, env, info, msg).map(customize_response::<C>)
-        };
-    Box::new(customized)
-}
-
-fn customize_response<C>(resp: Response<Empty>) -> Response<C>
-where
-    C: Clone + fmt::Debug + PartialEq + JsonSchema,
-{
-    Response::<C> {
-        messages: resp.messages.into_iter().map(customize_msg::<C>).collect(),
-        events: resp.events,
-        attributes: resp.attributes,
-        data: resp.data,
-    }
-}
-
-fn customize_msg<C>(msg: SubMsg<Empty>) -> SubMsg<C>
-where
-    C: Clone + fmt::Debug + PartialEq + JsonSchema,
-{
-    SubMsg {
-        msg: match msg.msg {
-            CosmosMsg::Wasm(wasm) => CosmosMsg::Wasm(wasm),
-            CosmosMsg::Bank(bank) => CosmosMsg::Bank(bank),
-            CosmosMsg::Staking(staking) => CosmosMsg::Staking(staking),
-            CosmosMsg::Custom(_) => unreachable!(),
-            #[cfg(feature = "stargate")]
-            CosmosMsg::Ibc(ibc) => CosmosMsg::Ibc(ibc),
-            #[cfg(feature = "stargate")]
-            CosmosMsg::Stargate { type_url, value } => CosmosMsg::Stargate { type_url, value },
-            _ => panic!("unknown message variant {:?}", msg),
-        },
-        id: msg.id,
-        gas_limit: msg.gas_limit,
-        reply_on: msg.reply_on,
-    }
-}
-
-impl<T1, T2, T3, E1, E2, E3, C, T4, E4, E5> Contract<C>
-    for ContractWrapper<T1, T2, T3, E1, E2, E3, C, T4, E4, E5>
-where
-    T1: DeserializeOwned,
-    T2: DeserializeOwned,
-    T3: DeserializeOwned,
-    T4: DeserializeOwned,
-    E1: ToString,
-    E2: ToString,
-    E3: ToString,
-    E4: ToString,
-    E5: ToString,
-    C: Clone + fmt::Debug + PartialEq + JsonSchema,
-{
-    fn execute(
-        &self,
-        deps: DepsMut,
-        env: Env,
-        info: MessageInfo,
-        msg: Vec<u8>,
-    ) -> Result<Response<C>, String> {
-        let msg: T1 = from_slice(&msg).map_err(|e| e.to_string())?;
-        let res = (self.execute_fn)(deps, env, info, msg);
-        res.map_err(|e| e.to_string())
-    }
-
-    fn instantiate(
-        &self,
-        deps: DepsMut,
-        env: Env,
-        info: MessageInfo,
-        msg: Vec<u8>,
-    ) -> Result<Response<C>, String> {
-        let msg: T2 = from_slice(&msg).map_err(|e| e.to_string())?;
-        let res = (self.instantiate_fn)(deps, env, info, msg);
-        res.map_err(|e| e.to_string())
-    }
-
-    // this returns an error if the contract doesn't implement sudo
-    fn sudo(&self, deps: DepsMut, env: Env, msg: Vec<u8>) -> Result<Response<C>, String> {
-        let msg: T4 = from_slice(&msg).map_err(|e| e.to_string())?;
-        let res = match &self.sudo_fn {
-            Some(sudo) => sudo(deps, env, msg),
-            None => return Err("sudo not implemented for contract".to_string()),
-        };
-        res.map_err(|e| e.to_string())
-    }
-
-    // this returns an error if the contract doesn't implement reply
-    fn reply(&self, deps: DepsMut, env: Env, reply_data: Reply) -> Result<Response<C>, String> {
-        let res = match &self.reply_fn {
-            Some(reply) => reply(deps, env, reply_data),
-            None => return Err("reply not implemented for contract".to_string()),
-        };
-        res.map_err(|e| e.to_string())
-    }
-
-    fn query(&self, deps: Deps, env: Env, msg: Vec<u8>) -> Result<Binary, String> {
-        let msg: T3 = from_slice(&msg).map_err(|e| e.to_string())?;
-        let res = (self.query_fn)(deps, env, msg);
-        res.map_err(|e| e.to_string())
-    }
-}
 
 struct ContractData {
     code_id: usize,
@@ -816,7 +548,7 @@ mod test {
 
     use crate::test_helpers::{contract_error, contract_payout, PayoutInitMessage, PayoutQueryMsg};
     use cosmwasm_std::testing::{mock_env, mock_info, MockApi, MockQuerier, MockStorage};
-    use cosmwasm_std::{coin, to_vec, BankMsg, BlockInfo, Coin, CosmosMsg, Empty};
+    use cosmwasm_std::{coin, from_slice, to_vec, BankMsg, BlockInfo, Coin, CosmosMsg, Empty};
 
     fn mock_router() -> WasmRouter<Empty> {
         let env = mock_env();
